Passando valores para decimal para uma maior precisao

+++ Cotacoes.API/Controllers/ConversaoController.cs
@@ -7,13 +7,13 @@ namespace Cotacoes.API.Controllers
     public class ConversaoController : Controller
     {
         [HttpGet("Dolar/{moeda}/{valor}")]
-        public Conversao ConverterParaDolar(string moeda, double valor)
+        public Conversao ConverterParaDolar(string moeda, decimal valor)
         {
             return ConversaoService.ConverterParaDolar(valor, moeda);
         }
 
         [HttpGet("Real/{moeda}/{valor}")]
-        public Conversao ConverterParaReal(string moeda, double valor)
+        public Conversao ConverterParaReal(string moeda, decimal valor)
         {
             return ConversaoService.ConverterParaReais(valor, moeda);
         }
