+++ Cotacoes.Model/conversao/ConversaoRepository.cs
@@ -5,7 +5,7 @@ namespace Cotacoes.Model
 {
     internal class ConversaoRepository : GenericRepository
     {
-        internal static Conversao ConverterParaReais(double Montante, string SiglaMoeda)
+        internal static Conversao ConverterParaReais(decimal Montante, string SiglaMoeda)
         {
             AtualizacoesCotacao.AtualizarCotacoes();
 
@@ -13,7 +13,7 @@ namespace Cotacoes.Model
             try
             {
                 AbrirConexao();
-                var valorConvercao = conexao.QueryFirst<Double>(sql, new { montante = Montante, sigla = SiglaMoeda.ToUpper() });
+                var valorConvercao = conexao.QueryFirst<decimal>(sql, new { montante = Montante, sigla = SiglaMoeda.ToUpper() });
                 FecharConexao();
 
                 return new Conversao(valorConvercao);
@@ -23,7 +23,7 @@ namespace Cotacoes.Model
                 return new Conversao(-1);
             }
         }
-        internal static Conversao ConverterParaDolar(double Montante, string SiglaMoeda)
+        internal static Conversao ConverterParaDolar(decimal Montante, string SiglaMoeda)
         {
             AtualizacoesCotacao.AtualizarCotacoes();
 
@@ -31,7 +31,7 @@ namespace Cotacoes.Model
             try
             {
                 AbrirConexao();
-                var valorConvercao = conexao.QueryFirst<Double>(sql, new { montante = Montante, sigla = SiglaMoeda.ToUpper() });
+                var valorConvercao = conexao.QueryFirst<decimal>(sql, new { montante = Montante, sigla = SiglaMoeda.ToUpper() });
                 FecharConexao();
 
                 return new Conversao(valorConvercao);
